Layout content components separation
